Build optimization (#953)
* Update supported browsers

* Update gitignore

* Update vue-loader and add dev pkgs

* Use babel plugin

* Update webpack config + measue build time

* Revert parallel true config

* Remove speed analysis

* Remove install husky on prepare

* Cleanup vue config

* fix: Re-introduce prepare command but prevent running husky install in builds

* Try turning parallel option back on after updating vue-loader

* fix: Looks like parallel = false is no longer required

* Move tailwind deps to dev deps

* Remove storybook references

* npm install forced
