+++ .browserslistrc
@@ -1,3 +1,3 @@
-> 1%
-last 2 versions
-not dead
+defaults
+not IE 11
+maintained node versions
