+++ .gitignore
@@ -4,6 +4,7 @@ node_modules
 /tailwind-static
 
 # local env files
+.env
 .env.local
 .env.*.local
