+++ README.md
@@ -53,11 +53,3 @@ The app is using [Tailwind](https://tailwindcss.com/) to configure base styles. 
 npm run tailwind-viewer
 ```
 Your browser should load the app at [http://localhost:3000](http://localhost:3000).
-
-### Component Library
-We are using [Storybook](https://storybook.js.org/) to document our commonly used components. You can browse the component library in development by running:
-
-```bash
-npm run storybook
-```
-Your browser should load the app at [http://localhost:6006](http://localhost:6006).
