+++ babel.config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  presets: ['@vue/cli-plugin-babel/preset']
+  presets: ['@vue/cli-plugin-babel/preset'],
+  plugins: ['@babel/plugin-proposal-numeric-separator']
 };
