+++ package.json
@@ -7,15 +7,13 @@
   },
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build --modern",
+    "build": "vue-cli-service build",
     "test:unit": "vue-cli-service test:unit --passWithNoTests",
     "lint": "vue-cli-service lint --no-fix --max-warnings 0",
     "import-abis": "cp ../core/deployments/artifacts/dockerParity/*.json ./src/utils/balancer/abi/.",
     "import-config": "tsc ./scripts/update-config.ts && node ./scripts/update-config.ts",
     "lint:fix": "vue-cli-service lint",
-    "prepare": "husky install",
-    "storybook": "start-storybook -p 6006",
-    "storybook:build": "build-storybook",
+    "prepare": "node ./prepare.js",
     "tailwind:build": "tailwind-config-viewer export ./tailwind-static",
     "tailwind:view": "tailwind-config-viewer -o"
   },
@@ -34,12 +32,10 @@
     "@portis/web3": "^3.0.10",
     "@sentry/browser": "^6.3.2",
     "@sentry/tracing": "^6.3.2",
-    "@tailwindcss/postcss7-compat": "^2.0.4",
     "@types/animejs": "^3.1.3",
     "@vueuse/core": "^4.3.4",
     "@walletconnect/web3-provider": "^1.5.2",
     "animejs": "^3.2.1",
-    "autoprefixer": "^9.8.6",
     "awesome-only-resolves-last-promise": "^1.0.3",
     "axios": "0.21.2",
     "bnc-sdk": "^3.5.0",
@@ -53,10 +49,7 @@
     "numeral": "^2.0.4",
     "paraswap": "^4.18.0",
     "parse-filepath": "1.0.2",
-    "postcss": "^7.0.35",
-    "postcss-nesting": "^7.0.1",
     "swiper": "^6.4.5",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.4",
     "vue": "^3.2.6",
     "vue-composable-tester": "^0.1.3",
     "vue-echarts": "^6.0.0",
@@ -72,12 +65,8 @@
   },
   "devDependencies": {
     "@babel/core": "^7.13.8",
+    "@babel/plugin-proposal-numeric-separator": "^7.14.5",
     "@sentry/webpack-plugin": "^1.15.1",
-    "@storybook/addon-actions": "^6.3.7",
-    "@storybook/addon-console": "^1.2.3",
-    "@storybook/addon-essentials": "^6.3.7",
-    "@storybook/addon-links": "^6.3.7",
-    "@storybook/vue3": "^6.3.7",
     "@types/jest": "^24.0.19",
     "@types/lodash": "^4.14.168",
     "@types/node": "^14.17.1",
@@ -96,18 +85,21 @@
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^7.0.0",
     "@vue/test-utils": "^2.0.0-alpha.1",
+    "autoprefixer": "^9.8.8",
     "babel-loader": "^8.2.2",
     "eslint": "^6.7.2",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-vue": "^7.16.0",
     "husky": "^6.0.0",
     "nock": "^13.1.1",
+    "postcss": "^7.0.39",
     "prettier": "^1.19.1",
     "promise-worker": "^2.0.1",
     "tailwind-config-viewer": "^1.5.0",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
     "typescript": "^4.3.5",
     "vue-cli-plugin-webpack-bundle-analyzer": "^2.0.0",
-    "vue-loader": "^16.5.0",
+    "vue-loader": "^16.8.1",
     "worker-loader": "^3.0.8"
   },
   "license": "MIT",
