+++ postcss.config.js
@@ -1,7 +1,3 @@
 module.exports = {
-  plugins: [
-    require('tailwindcss'),
-    require('postcss-nesting'),
-    require('autoprefixer')
-  ]
+  plugins: [require('tailwindcss'), require('autoprefixer')]
 };
